requirements.txt
streamlit==1.31.0 
pandas==2.1.0 
pyyaml==6.0.1 
python-docx==1.0.1 
